# libs/s3-integration/s3-integration-client/apidoc/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "FolderAPI",
      "description": "API to interact with folders"
    },
    {
      "name": "FileAPI",
      "description": "API to interact with files"
    }
  ],
  "paths": {
    "/file": {
      "get": {
        "tags": [
          "FileAPI"
        ],
        "description": "Creates a presigned URL to fetch the file specified in the parameter from the S3 storage",
        "operationId": "get",
        "parameters": [
          {
            "name": "pathToFile",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "expiresInMinutes",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FileAPI"
        ],
        "description": "Creates a presigned URL to overwrite the file specified in the parameter within the S3 storage",
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileDataDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FileAPI"
        ],
        "description": "Creates a presigned URL to store the file specified in the parameter within the S3 storage",
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileDataDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FileAPI"
        ],
        "description": "Creates a presigned URL to delete the file specified in the parameter from the S3 storage",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "pathToFile",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "expiresInMinutes",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "FileAPI"
        ],
        "description": "Updates the end of life attribute in the corresponding database entry for the file specified in the parameter",
        "operationId": "updateEndOfLife",
        "parameters": [
          {
            "name": "pathToFile",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "endOfLife",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/folder": {
      "get": {
        "tags": [
          "FolderAPI"
        ],
        "description": "Returns all file paths for the folder specified in the parameter",
        "operationId": "getAllFilesInFolderRecursively",
        "parameters": [
          {
            "name": "pathToFolder",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FilesInFolderDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FolderAPI"
        ],
        "description": "Deletes the folder specified in the parameter together with the corresponding database entry",
        "operationId": "delete",
        "parameters": [
          {
            "name": "pathToFolder",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FileDataDto": {
        "required": [
          "expiresInMinutes",
          "pathToFile"
        ],
        "type": "object",
        "properties": {
          "pathToFile": {
            "maxLength": 1024,
            "minLength": 0,
            "type": "string"
          },
          "expiresInMinutes": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "endOfLife": {
            "type": "string",
            "format": "date"
          }
        },
        "description": "File data for requesting a presigned url"
      },
      "PresignedUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "description": "Presigned file url"
      },
      "FilesInFolderDto": {
        "type": "object",
        "properties": {
          "pathToFiles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "File paths"
      }
    }
  }
}